requirements.txt
google-generativeai
chromadb
langchain
langchain-community
langchain-text-splitters
tqdm
python-dotenv
streamlit
docling
PyMuPDF
sentence_transformers
requests
trafilatura
unidecode
openai
underthesea
rank_bm25
nltk
pyvi